requirements.txt
fastapi
uvicorn
requests
python-dotenv
pyngrok
rich